Refactor worker stdin handling into focused modules
